+++ template_variables/src/epilot/sdk.py
@@ -17,7 +17,7 @@ class Epilot:
     sdk_configuration: SDKConfiguration
 
     def __init__(self,
-                 epilot_auth: Union[str,Callable[[], str]],
+                 security: Union[shared.Security,Callable[[], shared.Security]] = None,
                  server_idx: int = None,
                  server_url: str = None,
                  url_params: Dict[str, str] = None,
@@ -26,8 +26,8 @@ class Epilot:
                  ) -> None:
         """Instantiates the SDK configuring it with the provided parameters.
         
-        :param epilot_auth: The epilot_auth required for authentication
-        :type epilot_auth: Union[str,Callable[[], str]]
+        :param security: The security details required for authentication
+        :type security: Union[shared.Security,Callable[[], shared.Security]]
         :param server_idx: The index of the server to use for all operations
         :type server_idx: int
         :param server_url: The server URL to use for all operations
@@ -41,8 +41,6 @@ class Epilot:
         """
         if client is None:
             client = requests_http.Session()
-        
-        security = shared.Security(epilot_auth = epilot_auth)
         
         if server_url is not None:
             if url_params is not None: