+++ template_variables/src/epilot/sdkconfiguration.py
@@ -21,9 +21,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '2.1.4'
-    gen_version: str = '2.195.2'
-    user_agent: str = 'speakeasy-sdk/python 2.1.4 2.195.2 1.0.0 epilot-template-variables'
+    sdk_version: str = '2.1.5'
+    gen_version: str = '2.202.2'
+    user_agent: str = 'speakeasy-sdk/python 2.1.5 2.202.2 1.0.0 epilot-template-variables'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> Tuple[str, Dict[str, str]]: